working after small refactor
