+++ include/silo/query_engine/actions/aggregated.h
@@ -20,7 +20,7 @@ class Aggregated : public Action {
    void validateOrderByFields(const schema::TableSchema& schema) const override;
 
    [[nodiscard]] QueryResult execute(
-      const Database& database,
+      const std::shared_ptr<const storage::Table>& table,
       std::vector<CopyOnWriteBitmap> bitmap_filter
    ) const override;
