+++ include/silo/query_engine/actions/fasta.h
@@ -21,7 +21,7 @@ class Fasta : public Action {
    void validateOrderByFields(const schema::TableSchema& schema) const override;
 
    [[nodiscard]] QueryResult execute(
-      const Database& database,
+      const std::shared_ptr<const storage::Table>& table,
       std::vector<CopyOnWriteBitmap> bitmap_filter
    ) const override;
