+++ include/silo/query_engine/actions/insertions.h
@@ -13,10 +13,10 @@
 
 #include "silo/common/aa_symbols.h"
 #include "silo/common/nucleotide_symbols.h"
-#include "silo/database.h"
 #include "silo/query_engine/actions/action.h"
 #include "silo/query_engine/copy_on_write_bitmap.h"
 #include "silo/query_engine/query_result.h"
+#include "silo/storage/table.h"
 
 namespace silo::query_engine::actions {
 
@@ -50,7 +50,7 @@ class InsertionAggregation : public Action {
 
    std::unordered_map<std::string, InsertionAggregation<SymbolType>::PrefilteredBitmaps>
    validateFieldsAndPreFilterBitmaps(
-      const Database& database,
+      const std::shared_ptr<const storage::Table>& table,
       std::vector<CopyOnWriteBitmap>& bitmap_filter
    ) const;
 
@@ -60,7 +60,7 @@ class InsertionAggregation : public Action {
    void validateOrderByFields(const schema::TableSchema& schema) const override;
 
    [[nodiscard]] QueryResult execute(
-      const Database& database,
+      const std::shared_ptr<const storage::Table>& table,
       std::vector<CopyOnWriteBitmap> bitmap_filter
    ) const override;
