+++ include/silo/query_engine/actions/mutations.h
@@ -89,7 +89,7 @@ class Mutations : public Action {
    void validateOrderByFields(const schema::TableSchema& schema) const override;
 
    [[nodiscard]] QueryResult execute(
-      const Database& database,
+      const std::shared_ptr<const storage::Table>& table,
       std::vector<CopyOnWriteBitmap> bitmap_filter
    ) const override;
