+++ include/silo/query_engine/exec_node/legacy_result_producer.h
@@ -40,8 +40,9 @@ class LegacyResultProducer : public arrow::acero::ExecNode {
    LegacyResultProducer(
       arrow::acero::ExecPlan* plan,
       const std::vector<silo::schema::ColumnIdentifier>& columns,  // TODO const & ?
-      std::shared_ptr<Database> database,
-      std::shared_ptr<Query> query
+      const std::shared_ptr<const storage::Table>& table,
+      const std::vector<std::unique_ptr<filter::operators::Operator>>& partition_filter_operators,
+      const actions::Action* action
    );
 
    virtual const char* kind_name() const override { return "LegacyResultProducer"; }
