+++ src/silo/query_engine/actions/action.cpp
@@ -23,6 +23,8 @@
 #include "silo/query_engine/actions/mutations.h"
 #include "silo/query_engine/bad_request.h"
 #include "silo/query_engine/copy_on_write_bitmap.h"
+#include "silo/query_engine/exec_node/legacy_result_producer.h"
+#include "silo/query_engine/exec_node/ndjson_sink.h"
 #include "silo/query_engine/query_result.h"
 
 namespace silo::query_engine::actions {
@@ -107,10 +109,10 @@ void Action::setOrdering(
 static const size_t MATERIALIZATION_CUTOFF = 10000;
 
 QueryResult Action::executeAndOrder(
-   const Database& database,
+   const std::shared_ptr<const storage::Table>& table,
    std::vector<CopyOnWriteBitmap> bitmap_filter
 ) const {
-   validateOrderByFields(database.table->schema);
+   validateOrderByFields(table->schema);
 
    // Hacky solution to give the full feature set (randomization,
    // sorting) for small result sets, and streaming without those
@@ -122,7 +124,7 @@ QueryResult Action::executeAndOrder(
    }
    const bool is_large = num_rows > MATERIALIZATION_CUTOFF;
 
-   QueryResult result = execute(database, std::move(bitmap_filter));
+   QueryResult result = execute(table, std::move(bitmap_filter));
 
    if (result.isMaterialized() || !is_large) {
       SPDLOG_TRACE("materialized or small -> full featured sort, offset+limit");
@@ -282,6 +284,35 @@ std::vector<schema::ColumnIdentifier> columnNamesToFields(
       fields.emplace_back(column_name, column.value().type);
    }
    return fields;
+}
+
+QueryPlan Action::toQueryPlan(
+   const std::shared_ptr<const storage::Table>& table,
+   const std::vector<std::unique_ptr<filter::operators::Operator>>& partition_filter_operators,
+   std::ostream& output_stream
+) {
+   QueryPlan query_plan;
+   // TODO move to `toExecPlan` method
+   // but it will sometimes be more than one exec_node? select -> order -> limit
+   std::unique_ptr<arrow::acero::ExecNode> source_node;
+   source_node = std::make_unique<exec_node::LegacyResultProducer>(
+      query_plan.arrow_plan.get(),
+      getOutputSchema(table->schema),
+      table,
+      partition_filter_operators,
+      this
+   );
+
+   std::unique_ptr<arrow::acero::ExecNode> sink_node = std::make_unique<exec_node::NdjsonSink>(
+      query_plan.arrow_plan.get(), &output_stream, source_node.get()
+   );
+   // TODO make configurable std::make_unique<ArrowSinkNode>(query_plan.arrow_plan.get(),
+   // &output_stream, source_node.get());
+
+   query_plan.arrow_plan->AddNode(std::move(source_node));
+   query_plan.arrow_plan->AddNode(std::move(sink_node));
+
+   return query_plan;
 }
 
 }  // namespace silo::query_engine::actions
