+++ src/silo/query_engine/actions/fasta_aligned.cpp
@@ -22,14 +22,19 @@
 #include "silo/query_engine/actions/action.h"
 #include "silo/query_engine/bad_request.h"
 #include "silo/query_engine/copy_on_write_bitmap.h"
+#include "silo/query_engine/exec_node/ndjson_sink.h"
 #include "silo/query_engine/exec_node/select.h"
 #include "silo/query_engine/query_result.h"
 #include "silo/storage/column/sequence_column.h"
 
 namespace silo::query_engine::actions {
 
-FastaAligned::FastaAligned(std::vector<std::string>&& sequence_names, std::vector<std::string>&& additional_fields)
-    : sequence_names(sequence_names), additional_fields(additional_fields) {}
+FastaAligned::FastaAligned(
+   std::vector<std::string>&& sequence_names,
+   std::vector<std::string>&& additional_fields
+)
+    : sequence_names(sequence_names),
+      additional_fields(additional_fields) {}
 
 void FastaAligned::validateOrderByFields(const schema::TableSchema& schema) const {
    const std::string& primary_key_field = schema.primary_key.name;
@@ -49,7 +54,7 @@ void FastaAligned::validateOrderByFields(const schema::TableSchema& schema) cons
 }
 
 QueryResult FastaAligned::execute(
-   const Database& database,
+   const std::shared_ptr<const storage::Table>& table,
    std::vector<CopyOnWriteBitmap> bitmap_filter
 ) const {
    SILO_PANIC("Legacy execute called on action already migrated action. Programming error.");
@@ -73,20 +78,20 @@ std::vector<schema::ColumnIdentifier> FastaAligned::getOutputSchema(
    return fields;
 }
 
-QueryPlan FastaAligned::toQueryPlan() {
+QueryPlan FastaAligned::toQueryPlan(
+   const std::shared_ptr<const storage::Table>& table,
+   const std::vector<std::unique_ptr<filter::operators::Operator>>& partition_filter_operators,
+   std::ostream& output_stream
+) {
    QueryPlan query_plan;
-   auto table_schema = database->schema.tables.at(schema::TableName::getDefault());
    // TODO move to `toExecPlan` method
    // but it will sometimes be more than one exec_node? select -> order -> limit
    std::unique_ptr<arrow::acero::ExecNode> source_node;
    source_node = std::make_unique<exec_node::Select>(
-      query_plan.arrow_plan.get(),
-      getOutputSchema(table_schema),
-      ,
-      database
+      query_plan.arrow_plan.get(), getOutputSchema(table->schema), partition_filter_operators, table
    );
 
-   std::unique_ptr<arrow::acero::ExecNode> sink_node = std::make_unique<NdjsonSinkNode>(
+   std::unique_ptr<arrow::acero::ExecNode> sink_node = std::make_unique<exec_node::NdjsonSink>(
       query_plan.arrow_plan.get(), &output_stream, source_node.get()
    );
    // TODO make configurable std::make_unique<ArrowSinkNode>(query_plan.arrow_plan.get(),
