+++ src/silo/query_engine/exec_node/sort.cpp
@@ -2,9 +2,9 @@
 
 #include "silo/query_engine/exec_node/arrow_util.h"
 //
-//namespace silo::query_engine::exec_node {
+// namespace silo::query_engine::exec_node {
 //
-//Sort::Sort(
+// Sort::Sort(
 //   arrow::acero::ExecPlan* plan,
 //   const std::vector<silo::schema::ColumnIdentifier>& columns,  // TODO const & ?
 //   std::shared_ptr<Database> database,
@@ -18,13 +18,13 @@
 //   prepareOutputArrays();
 //}
 //
-//void LegacyResultProducer::prepareOutputArrays() {
+// void LegacyResultProducer::prepareOutputArrays() {
 //   for (auto& field : output_schema_.get()->fields()) {
 //      arrays.emplace_back(field->type());
 //   }
 //}
 //
-//arrow::Status LegacyResultProducer::flushOutput() {
+// arrow::Status LegacyResultProducer::flushOutput() {
 //   std::vector<arrow::Datum> data;
 //   for (auto& array : arrays) {
 //      data.push_back((std::move(array)).toDatum());
@@ -36,42 +36,41 @@
 //}
 //
 //// TODO make configurable
-//static constexpr size_t MATERIALIZATION_CUTOFF = 50000;
+// static constexpr size_t MATERIALIZATION_CUTOFF = 50000;
 //
-//arrow::Status LegacyResultProducer::produce() {
-//   size_t num_rows = 0;
-//   std::optional<QueryResultEntry> row;
-//   while ((row = query_result.next())) {
-//      ++num_rows;
-//      for (size_t field_idx = 0; field_idx < field_names.size(); ++field_idx) {
-//         const auto field_name = field_names.at(field_idx);
-//         const common::JsonValueType& field_value = row.value().fields.at(*field_name);
+// arrow::Status LegacyResultProducer::produce() {
+//    size_t num_rows = 0;
+//    std::optional<QueryResultEntry> row;
+//    while ((row = query_result.next())) {
+//       ++num_rows;
+//       for (size_t field_idx = 0; field_idx < field_names.size(); ++field_idx) {
+//          const auto field_name = field_names.at(field_idx);
+//          const common::JsonValueType& field_value = row.value().fields.at(*field_name);
 //
-//         auto status = arrays.at(field_idx).insert(field_value);
-//         if (status.IsCapacityError()) {
-//            throw std::runtime_error(fmt::format(
-//               "Response size too large. Materializing {} rows required more than allowed {} bytes",
-//               MATERIALIZATION_CUTOFF,
-//               INT32_MAX
-//            ));
-//         }
-//         ARROW_RETURN_NOT_OK(status);
-//      }
+//          auto status = arrays.at(field_idx).insert(field_value);
+//          if (status.IsCapacityError()) {
+//             throw std::runtime_error(fmt::format(
+//                "Response size too large. Materializing {} rows required more than allowed {}
+//                bytes", MATERIALIZATION_CUTOFF, INT32_MAX
+//             ));
+//          }
+//          ARROW_RETURN_NOT_OK(status);
+//       }
 //
-//      if (num_rows > MATERIALIZATION_CUTOFF) {
-//         ARROW_RETURN_NOT_OK(flushOutput());
-//      }
-//   }
-//   ARROW_RETURN_NOT_OK(flushOutput());
-//   return arrow::Status::OK();
-//}
+//       if (num_rows > MATERIALIZATION_CUTOFF) {
+//          ARROW_RETURN_NOT_OK(flushOutput());
+//       }
+//    }
+//    ARROW_RETURN_NOT_OK(flushOutput());
+//    return arrow::Status::OK();
+// }
 //
-//arrow::Status LegacyResultProducer::StartProducing() {
-//   return produce();
-//}
+// arrow::Status LegacyResultProducer::StartProducing() {
+//    return produce();
+// }
 //
-//arrow::Status LegacyResultProducer::StopProducing() {
-//   return arrow::Status::OK();
-//}
+// arrow::Status LegacyResultProducer::StopProducing() {
+//    return arrow::Status::OK();
+// }
 //
-//}  // namespace silo::query_engine::exec_node
+// }  // namespace silo::query_engine::exec_node
