+++ src/silo/query_engine/query.cpp
@@ -2,11 +2,13 @@
 
 #include <string>
 
+#include <spdlog/spdlog.h>
 #include <nlohmann/json.hpp>
 
 #include "silo/query_engine/actions/action.h"
 #include "silo/query_engine/bad_request.h"
 #include "silo/query_engine/filter/expressions/expression.h"
+#include "silo/query_engine/filter/operators/operator.h"
 
 namespace silo::query_engine {
 
@@ -26,6 +28,24 @@ std::shared_ptr<Query> Query::parseQuery(const std::string& query_string) {
    } catch (const nlohmann::json::exception& ex) {
       throw BadRequest("The query was not a valid JSON: " + std::string(ex.what()));
    }
+}
+
+QueryPlan Query::toQueryPlan(std::shared_ptr<Database> database, std::ostream& output_stream)
+   const {
+   SPDLOG_DEBUG("Parsed filter: {}", filter->toString());
+
+   std::vector<std::unique_ptr<filter::operators::Operator>> partition_filter_operators;
+   for (size_t partition_index = 0; partition_index < database->table->getNumberOfPartitions();
+        partition_index++) {
+      partition_filter_operators.emplace_back(filter->compile(
+         *database,
+         database->table->getPartition(partition_index),
+         filter::expressions::Expression::AmbiguityMode::NONE
+      ));
+      SPDLOG_DEBUG("Simplified query: ", partition_filter_operators.back()->toString());
+   };
+
+   return action->toQueryPlan(database->table, partition_filter_operators, output_stream);
 }
 
 }  // namespace silo::query_engine
